+++ src/extensions.js
@@ -32,11 +32,11 @@ IDE_Morph, CamSnapshotDialogMorph, SoundRecorderDialogMorph, isSnapObject, nop,
 Color, Process, contains, localize, SnapTranslator, isString, detect, Point,
 SVG_Costume, newCanvas, WatcherMorph, BlockMorph, HatBlockMorph, invoke, isNil,
 BigUint64Array, DeviceOrientationEvent, DialogBoxMorph, Animation, TableMorph,
-TableFrameMorph, console, Morph*/
+TableFrameMorph, console, Morph, MenuMorph*/
 
 /*jshint esversion: 11, bitwise: false*/
 
-modules.extensions = '2026-February-13';
+modules.extensions = '2026-March-04';
 
 // Global stuff
 
@@ -1416,6 +1416,13 @@ SnapExtensions.primitives.set(
     'cst_morph(cst)',
     function (costume, proc) {
         var m = new Morph(),
+            imageView = () => new DialogBoxMorph().inform(
+                'Image view',
+                null,
+                this.world(),
+                costume.contents,
+                true // enable exporting
+            ),
             img;
         proc.assertType(costume, 'costume');
         img = costume.contents;
@@ -1423,7 +1430,34 @@ SnapExtensions.primitives.set(
         m.bounds.setWidth(img.width);
         m.bounds.setHeight(img.height);
         m.cachedImage = img;
+        m.isCustomSwatch = true;
+
+        m.mouseDoubleClick = imageView;
+
+        // support exporting costumes directly from speech balloons:
+        m.userMenu = function () {
+            var menu = new MenuMorph(this),
+                ide = this.parentThatIsA(IDE_Morph)||
+                    this.world().childThatIsA(IDE_Morph);
+
+            if (ide.isAppMode) {return; }
+            menu.addItem('open in dialog...', imageView);
+            menu.addItem(
+                'export',
+                () => ide.saveCanvasAs(img, costume.name || 'image')
+            );
+            return menu;
+        };
+
         return m;
+    }
+);
+
+SnapExtensions.primitives.set(
+    'cst_shrink-wrap(cst)',
+    function (costume) {
+        costume.shrinkWrap();
+        return costume;
     }
 );
 
@@ -1747,6 +1781,22 @@ SnapExtensions.primitives.set(
         data.map(eachRow => dict[eachRow.at(1)] = eachRow.at(2));
         SnapTranslator.dict[SnapTranslator.language] = dict;
         ide.reflectLanguage(SnapTranslator.language);
+    }
+);
+
+SnapExtensions.primitives.set(
+    'ide_switch_to_palette(category)',
+    function (category) {
+        var ide = this.parentThatIsA(IDE_Morph);
+        if (ide.scene.unifiedPalette) {
+            ide.scrollPaletteToCategory(category);
+        } else {
+            ide.currentCategory = category;
+            ide.categories.buttons.forEach(each =>
+                each.refresh()
+            );
+            ide.refreshPalette(true);
+        }
     }
 );