+++ HISTORY.md
@@ -31,16 +31,26 @@
     * custom data types (ADTs)
         * include a ring entry named underscore + morph in data to specify a dynamic view
         * new "cst_morph(cst)" extension for specifying custom ADT visualizations
+        * include an entry named underscore + field in the data to specify a custom data type
+        * integrated type-inferral for user defined data types
+        * new ADT input slot for blocks
+    * enforcing data types in input slots
+        * new optional "reports" type declaration entry for custom reporters
+        * new "enforce types" option for all custom blocks: only lets users drop reporters into input slots whose return type matches that of the slot
+        * new general "enforce input types" preference setting for scenes / puzzles / microworlds
     * domain-specific languages (DSL) support
         * support for dynamic dropdown menus and read-only settings in variadic input slots
         * support for variadic upvars in custom blocks
         * new 'parameter' custom block input type, same as 'upvar', but doesn't add a script variable when evaluated, supports variadicity
         * support for dynamically setting the contents of expanded variadic upvars and input slots by a user script in the block definition
         * support for user-scriptable rename-menus in upvars, including variadic ones (a single menu is repeated in each non-unevaluated slot, a list of menu-lists whose first item is an empty list gets repeated across all subslots)
+        * new "360° angles" dial widget for mathematical bearings (zero is East, counterclockwise) instead of compass bearings
+        * new "number unevaluated" special input slot for custom blocks
     * files
         * new general "dta_export(data, name, type)" extension (for csv, json, etc.)
         * new general "dta_import(raw?)" extension (for text files, csv, json etc.)
     * ui
+        * change name and color of custom categories
         * new "blocks" and "speaker" symbols, icons for the sprite-editor tabs
         * new "hide empty categories" setting
         * added a black-gray-white palette to the bottom of the color picker for color input slots
@@ -51,10 +61,14 @@
         * new reporter version of "let" in the variables declaration extension
         * new "... times ..." reporter in the neural networks library for quantifying results
         * new blocks to load costumes and sounds from a url in the web-services library, thanks, @ego-lay-atman-bay!
+        * new "cst_shrink-wrap(cst)" extension
+        * new "ide_switch_to_palette(category)" extension
     * new help screens
         * new help screens for "combinations" and
         * "pipe" reporters, thanks, Brian and gang!
 * **Notable Changes:**
+    * data types
+        * enhanced "is ... a ...?" predicate to support dependent data types and user defined structs (ADTs)
     * ui
         * keep result balloons "attached" to their originating block / script when scrolling
         * added icons to the IDE tabs (scripts, costumes, sounds)
@@ -65,6 +79,8 @@
         * sprite icons no longer blur when the sprite size is reduced or the zoom level increases
         * resizing the browser / window in presentation ("app") mode scales the stage smoothly
         * removed "stage selected, no motion primitives" text from the palette to support sprite-less microworlds
+    * blocks
+        * removed landscape orientation of text-input slots
     * meta-programming
         * silently handle missing variable references in user-scripted dropdowns and expansion subslots by returning an empty list instead
         * removed "static" tag from the block-attribute getter
@@ -96,6 +112,69 @@
     * Chinese, thank you, @Ayist14 !
     * Catalan, thanks, Joan!
     * German
+
+### 2026-03-13
+* blocs, objects: made text-input slots type aware and removed landscape orientation
+
+### 2026-03-11
+* tweaked slot type matching
+
+### 2026-03-10
+* blocks: changed ADT input slots evaluation to return an empty list
+* updated the tables extension with new data type enforcements
+* updated the shapes extension with new data type enforcements
+* gui: experimental hidden "enforce input types" preference setting (per session)
+* gui, scenes, store: new "enforce input types" setting for scenes and puzzles / microworlds
+* updated imbw blumen microworld with the new "enforce input types" option
+* blocks, byob: new "number unevaluated" special input slot for custom blocks
+* updated shapes extension with new "number unevaluated" special input slots
+
+### 2026-03-09
+* objects, blocks, byob: "enforce types" option for custom blocks: only lets users drop reporters into input slots whose return type matches that of the slot
+* updated the neural networks extension with new data type enforcements
+* updated the plot bars extension with new data type enforcements
+
+### 2026-03-05
+* updated the tutorial extension with a new "select editor category" command
+
+### 2026-03-04
+* extensions: new "ide_switch_to_palette(category)" extension
+
+### 2026-03-03
+* extensions, widgets: enable exporting images from image-view dialogs
+* shapes module update
+
+### 2026-03-02
+* objects, blocks, widgets, gui: change name and color of custom categories
+* extensions, widgets: "open in dialog..." menu option for costume-morphs, double-click to pop up images in dialog
+
+### 2026-02-28
+* added  (equilateral) triangles to the shapes library
+
+### 2026-02-27
+* threads: metaprogramming support for the ADT slot
+* blocks, extensions: enable exporting custom graphical ADT representations from result balloons and speech bubbles
+* British-English translation update
+* updated neural networks, tables and shapes extensions with new ADT input slots
+
+### 2026-02-26
+* blocks, byob, store: new ADT type input slot
+* blocks, byob, store: added optional type annotations and strict typing to blocks and custom blocks
+
+### 2026-02-25
+* updated neural networks, tables and shapes libraries with new type annotation format (underscore-type string)
+* blocks, threads: changed "is ... a ... ?" predicate to support user defined structs (ADTs) and dependent data types
+
+### 2026-02-24
+* extensions: new "cst_shrink-wrap(cst)" extension
+* objects: speed-up for generated costumes, but now need to explicitly shrink-wrap them
+* updated the neural networks library with faster previews for ADTs
+* updated shapes library with math angles for rotation (instead of compass bearings) and preview speed-up
+* gui: fixed the palette handle when hiding empty categories
+* updated the shapes library with a new "scale ... by x: ... y: ..." block
+
+### 2026-02-21
+* morphic, blocks, byob: new "360° angles" dial widget for mathematical bearings (zero is East, counterclockwise) instead of compass bearings
 
 ### 2026-02-18
 * gui: added "Blocks only" setting to the "looks" menu - hides the stage and all sprite editing panes for custom "functions-first" microworlds